+++ src/components/MainSection/index.tsx
@@ -28,7 +28,7 @@ const categories: ICategoryBox[] = [
     category: '대처 가이드라인',
     example: 'EX.) 보험은 있어요',
     img: info,
-    path: '/guideline',
+    path: '/survey',
   },
   {
     category: '집이 경매에 넘어갔다면',
